spotter-back-end/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
geographiclib==2.1
geopy==2.4.0
idna==3.10
mysqlclient==2.2.4
openrouteservice==2.3.3
Pillow==11.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.6.0
